+++ src/wurf/config_file.py
@@ -9,33 +9,32 @@ except ImportError:
 
 import os
 
-
-CONFIG_FILE = os.path.expanduser('~/.wurf_config')
+LOCAL_CONFIG_FILE = os.path.abspath('.wurf_config')
+USER_CONFIG_FILE = os.path.abspath(os.path.expanduser('~/.wurf_config'))
 
 
 class ConfigFile(object):
 
     def __init__(self, ctx):
-        
         self.default_resolve_path = None
-        self.default_symlinks_path = None
         self.ctx = ctx
-        if not os.path.isfile(CONFIG_FILE):
+        if os.path.isfile(LOCAL_CONFIG_FILE):
+            config_file = LOCAL_CONFIG_FILE
+        elif os.path.isfile(USER_CONFIG_FILE):
+            config_file = USER_CONFIG_FILE
+        else:
+            # No config file found
             return
-        
+
         self.ctx.start_msg("Using config file")
         config = configparser.ConfigParser()
         try:
-            config.read(CONFIG_FILE)
+            config.read(config_file)
         except configparser.Error as e:
             self.ctx.end_msg("ERROR: " + e.message, color='RED')
             return
 
-        self.ctx.end_msg(CONFIG_FILE)
+        self.ctx.end_msg(config_file)
         if config.has_option('DEFAULT', 'resolve_path'):
             self.default_resolve_path = config.get('DEFAULT', 'resolve_path')
-            self.ctx.msg("resolve_path", self.default_resolve_path)
-
-        if config.has_option('DEFAULT', 'symlinks_path'):
-            self.default_symlinks_path = config.get('DEFAULT', 'symlinks_path')
-            self.ctx.msg("symlinks_path", self.default_symlinks_path)
+            self.ctx.msg("Default resolve path", self.default_resolve_path)