+++ test/add_dependency/test_add_dependency.py
@@ -189,7 +189,8 @@ def run_commands(app_dir, git_dir):
     app_dir.run(['python', 'waf', '--help'], env=env)
 
     # We should be able to use --foo_magic_option that is defined in 'foo'
-    app_dir.run(['python', 'waf', 'configure', '-v', '--foo_magic_option=xyz'])
+    app_dir.run(['python', 'waf', 'configure', '-v', '--foo_magic_option=xyz',
+                 '--resolve_path', 'resolved_dependencies'])
 
     # After configure, the help text should include the description of
     # --foo_magic_option (defined in the 'foo' wscript)
@@ -213,7 +214,8 @@ def run_commands(app_dir, git_dir):
         app_dir.path(), 'resolve_symlinks', 'baz', 'libqux'))
 
     app_dir.run(['python', 'waf', 'build', '-v'])
-    app_dir.run(['python', 'waf', 'configure', '-v', '--fast_resolve'])
+    app_dir.run(['python', 'waf', 'configure', '-v', '--fast_resolve',
+                 '--resolve_path', 'resolved_dependencies'])
     app_dir.run(['python', 'waf', 'build', '-v'])
 
     # Test the zones print
@@ -224,7 +226,8 @@ def run_commands(app_dir, git_dir):
     assert r.stdout.match('* resolve recurse bar *')
 
     # Try the use checkout
-    app_dir.run(['python', 'waf', 'configure', '-v', '--baz_checkout=4.0.0'])
+    app_dir.run(['python', 'waf', 'configure', '-v', '--baz_checkout=4.0.0',
+                 '--resolve_path', 'resolved_dependencies'])
     app_dir.run(['python', 'waf', 'build', '-v'])
 
     # Lets remove the resolved dependencies
@@ -232,7 +235,8 @@ def run_commands(app_dir, git_dir):
     resolve_dir.rmdir()
 
     # Test the --lock_versions options
-    app_dir.run(['python', 'waf', 'configure', '-v', '--lock_versions'])
+    app_dir.run(['python', 'waf', 'configure', '-v', '--lock_versions',
+                 '--resolve_path', 'resolved_dependencies'])
     assert app_dir.contains_file('lock_resolve.json')
 
     # The symlinks should be available to all dependencies
@@ -253,7 +257,8 @@ def run_commands(app_dir, git_dir):
     resolve_dir.rmdir()
 
     # This configure should happen from the lock
-    app_dir.run(['python', 'waf', 'configure', '-v'])
+    app_dir.run(['python', 'waf', 'configure', '-v',
+                 '--resolve_path', 'resolved_dependencies'])
 
     assert app_dir.contains_dir('resolve_symlinks', 'foo')
     assert app_dir.contains_dir('resolve_symlinks', 'baz')
@@ -285,7 +290,8 @@ def run_commands(app_dir, git_dir):
 
     # Test the --lock_paths options
     app_dir.run(['python', 'waf', 'configure', '-v', '--lock_paths',
-                 '--resolve_path', 'locked'])
+                 '--resolve_path', 'locked',
+                 '--resolve_path', 'resolved_dependencies'])
 
     assert app_dir.contains_dir('resolve_symlinks', 'foo')
     assert app_dir.contains_dir('resolve_symlinks', 'baz')
@@ -297,7 +303,8 @@ def run_commands(app_dir, git_dir):
     # This configure should happen from the lock
     # Now we can delete the git folders - as we should be able to configure
     # from the frozen dependencies
-    app_dir.run(['python', 'waf', 'configure', '-v'])
+    app_dir.run(['python', 'waf', 'configure', '-v',
+                 '--resolve_path', 'resolved_dependencies'])
 
     assert app_dir.contains_dir('resolve_symlinks', 'foo')
     assert app_dir.contains_dir('resolve_symlinks', 'baz')
@@ -402,7 +409,7 @@ def test_add_dependency_path(testdirectory):
     baz_dir = mkdir_libbaz(directory=path_test, qux_dir=qux_dir)
 
     app_dir.run(['python', 'waf', 'configure', '-v', '--baz_path={}'.format(
-                baz_dir.path())])
+                baz_dir.path()), '--resolve_path', 'resolved_dependencies'])
 
     # The symlinks should be available to all dependencies
     assert os.path.exists(os.path.join(
@@ -413,7 +420,8 @@ def test_add_dependency_path(testdirectory):
         app_dir.path(), 'resolve_symlinks', 'bar'))
 
     app_dir.run(['python', 'waf', 'build', '-v'])
-    app_dir.run(['python', 'waf', 'configure', '-v', '--fast_resolve'])
+    app_dir.run(['python', 'waf', 'configure', '-v', '--fast_resolve',
+                 '--resolve_path', 'resolved_dependencies'])
 
 
 def test_create_standalone_archive(testdirectory):
@@ -439,7 +447,8 @@ def test_create_standalone_archive(testdirectory):
     with open(json_path, 'w') as json_file:
         json.dump(clone_path, json_file)
 
-    app_dir.run(['python', 'waf', 'configure', '-v', '--lock_paths'])
+    app_dir.run(['python', 'waf', 'configure', '-v', '--lock_paths',
+                 '--resolve_path', 'resolved_dependencies'])
     app_dir.run(['python', 'waf', '-v', 'standalone'])
     assert app_dir.contains_file('test_add_dependency-1.0.0.zip')
 
@@ -475,7 +484,8 @@ def test_override_json(testdirectory):
         json.dump(clone_path, json_file)
 
     # Try the use checkout
-    app_dir.run(['python', 'waf', 'configure', '-v'])
+    app_dir.run(['python', 'waf', 'configure', '-v',
+                 '--resolve_path', 'resolved_dependencies'])
     app_dir.run(['python', 'waf', 'build', '-v'])
 
     resolve_dir = app_dir.join('resolved_dependencies')